// Marketplace/src/types.ts
export interface User {
  id: string;
  phone: string;
  first_name: string;
  last_name: string;
  gender: 'erkak' | 'ayol';
  region_id: number;
  district_id: number;
  mfy_id: number;
  birth_date: string;
}
// ...
export interface Region {
  id: number;
  name: string;
}
// ...
export interface ActivityType {
  id: number;
  name: string;
}

export interface CommentTemplate {
  id: number;
  comment: string;
  sort_order?: number;
}

export interface PartnerRequestPayload {
  company_name: string;
  inn: string;
  mfo: string;
  account_number: string;
  activity_type_id: number;
  region_id: number;
  district_id: number;
  mfy_id: number;
  phone: string;
}

export interface PartnerRequest {
  id: number;
  company_name: string;
  inn: string;
  mfo: string;
  account_number: string;
  activity_type_id: number;
  activity_type_name?: string;
  region_id: number;
  district_id: number;
  mfy_id: number;
  phone: string;
  status?: string;
  created_at?: string;
}